Stop dumping all possible messages into 'messages' stream.
This commit introduces the 'private' stream, as well as unique
'search/*' streams that get created for every new search. This allows
the model to be filtered by the stream column, in such a way that it's
possible to show search results separately from the rest of the
messages (previously, if you performed a search, there would be no way
to distinguish search results from any other message in the 'messages' stream).