Fix SharedModel sorting, thanks to mhr3.
Apparently you can't cmp() a GLib.Variant the way you would expect to
do so with a string, so you have to .get_string() from the Variant
first.

This commit also adds tests for _cmp and _cmp_date, and also fixes up
a bunch of other tests which were broken by the model data being sorted.