Fix some AccountManager methods. Fixes lp:1069509
Ok, so I reviewed this in much more depth than previously. This
commit:

* Unifies the behavior of disabling and deleting accounts. Now, if
  you're in UOA, and you flick an account to 'OFF', it's messages will
  be removed from the model. Similarly, if you outright delete an
  account, the associated messages are also deleted. Currently we do
  not make any distinction between a "disabled" and a "deleted"
  account. In both cases, messages are removed from the model, and no
  attempts are made to redownload messages from that account, until it
  is re-enabled later.

* Removes the "refresh callback" from the AccountManager, because that
  was stupid. It turns out, when an account is enabled, you only need
  to call 'receive' on that one account, not on all of them together.
  So this is a nice little efficiency win.

* Updates all the tests to the new API changes.

I tested this thoroughly with friends-service.sh, using UOA to enable,
disable, delete, and recreate accounts, and it works great.

The best part is, thanks to my previous work on using
Model.insert_sorted instead of Model.append, deleting and re-adding
messages to the model doesn't change the order of those messages in
the Model.