Drop the 'comments' 'as' field from the model schema.
Facebook's replies have been promoted to top-level comments in the
model. Instead of appearing in the 'messages' stream, there is now a
new set of streams in the format of 'reply_to/*' where * is the
message_id that is being replied to.

This commit also fixes up all the associated test breakages, with some
new test coverage of the new behavior.