Generalize Dispatcher.Do.
It is now smarter about not passing in blank message_id kwargs to
methods that don't support it. Unfortunately DBus does not allow
arguments to be empty, so any DBus clients to this library are going
to be forced to pass in empty strings for account_id and message_id
when they are not appropriate. So now, if account_id is passed in an
empty string, then Do iterates over all accounts. and if message_id is
blank, Do does not attempt to pass it in. This gives us a bit more
flexibility in our method calls.

Although it still doesn't make it possible to do searches, really, it
does make it possible to call 'home' or 'receive' or whatever without
arguments.