Standardize Flickr protocol.
Flickr was one of the first protocols written and thus it had a lot of
nonstandard code from before we had a good understanding of the
overall architecture. So I've ripped out _get_nsid in favor of the
more standard _get_access_token, and adjusted everything else
(including tests) to accomodate this change.

Also, I've implemented the sender_id column and Avatar logic in line
with how Facebook/Twitter are currently doing it, so that's a nice
little increase in consistency for us.