Add a test case for the Uploader.
I've added a 'mirror' option to the WSGI test environment that just
spits back at you whatever you send to it, including uploaded data.
Then we do an upload to it, and test that the raw data given back to
us looks like a valid HTTP upload (without checking each individual
byte of the uploaded file, that would be huge).