Rip out venv/check_all logic.
This code was largely a holdover from the python2 days, before we had
ported to Dee.SharedModels, because back then we had basically
reimplmented Dee.SharedModel from scratch, and needed a huge amount of
testing framework overhead in order to test it; Now that we're using
Dee.SharedModel instead, we got to drop almost all of the code that
needed to be tested live via dbus, leaving very few tests running
inside this enormous framework.

One of the last pieces of this puzzle was the well-mocked test case
for the dispatcher, which was previously only believed to be testable
through a live dbus session. But now that we've just mocked dbus right
out in order to achieve simpler, lighter unit tests, this whole dbus
testing framework is largely redundant.

On top of all this, it's currently broken with mysterious error
messages, I'm just ripping the whole thing out and moving the last
useful test into test_dispatcher.py instead.