Catch exceptions and call the failure callback instead of (uselessly) raising them.
Our threading infrastructure makes exceptions really useless because
they get silently swallowed and then logged -- calling code has no
real way of checking for error conditions short of parsing the log
file in real time. So this commit (at least for facebook uploads)
catches all exceptions and calls the failure callback.

At this point I am not aware of any other possible exceptions that
could be raised, that we are not already catching and calling the
failure callback. Ie, there are no conditions known to me in which an
exception will be logged and ignored instead of the failure callback
being called.

This should be considered the model for how to implement async APIs
throughout the rest of lp:friends, and any other methods needing to
indicate error conditions back to it's calling code will have to
mimick this design.