Fully generalize all the async code so that any protocol operation can utilize it without needing any modifications.
This commit also adds the async API onto Dispatcher.Do, SendMessage,
and SendReply, and re-unifies UploadAsync into Upload. The callbacks
are fully optional, so all previous code should be unaffected and
continue working, but you now have the option of adding callbacks to
be notified when the operations complete.

This commit also adds test coverage to _OperationThread, which I
discovered was untested.