Get better about raising exceptions to indicate error conditions.
Now that we have a viable way of indicating error conditions to our
calling code, with the failure callback, it no longer makes sense to
log & ignore error conditions. So, this commit changes most (if not
all...) of the calls to 'log.error' with calls to 'raise FriendsError'
instead.

These exceptions still end up being logged thanks to
_OperationThread's traceback logging logic, and also prevent the
success callback from being called.

I think we're in a pretty good state here, although we'll have to keep
an eye on this and make sure that the success callback never gets
called when there was actually an error. If we find such a case, we'll
need to fix it by identifying the error condition and then ensuring
that it raises an appropriate exception.