Move some authorization-related exception raising logic out of Base class and into the Authentication class, where it makes much more sense to exist.
Previously the Authentication class was just returning None to
indicate various error conditions, and then we were checking for that
and raising exceptions in not one but three different places in Base.
This simplifies the code without losing any functionality, and improves
the meaningfulness of exception messages.

This commit also updates the tests to accomodate these changes,
dropping some obsolete tests from the protocol test suites and
replacing them with a better, centralized test inside
test_authentication.py instead.