Reduce memory usage by storing only row index ints rather than DeeModelRowIter objects, which wastefully contain references to many, many, many unnecessary python objects.
In my limited testing, this change dropped memory usage by ~10MBs for
a DeeModel containing 1660 rows, and I confirmed with the objgraph
module that triggering a refresh no longer results in any additional
python objects being kept around in memory -- in theory we have
abolished all memory leaks! In practise however, C code that we run
through GObject Introspection is free to leak as it pleases and this
tool would not detect it.