Disallow posting to disabled accounts. (Closes LP: 1103717)
This one was a bit more involved than I had originally anticipated. It
turns out, revision 113 where we drop this '#/protocol' business and
just use the libaccounts account # directly was incomplete. Our
AccountManager class was still constructing this id string, and it was
inconsistent with the way the ids were being handled everywhere else
in the code.

As a result of this, we experienced a regression such that disabled
accounts weren't being removed from memory correctly, and thus calling
SendMessage began sending messages to accounts that were, in theory,
deleted. This is now fixed.

In addition to this, FakeAccount mock class has been updated to
reflect an id string that is more like what actually gets used in
practise, and many, many, many tests have been updated to reflect this
new value.

While troubleshooting this, I also discovered a new issue that was
introduced in revision 102: when we reduced memory usage by storing
only row index ints rather than DeeModelRowIter objects. The problem
was that when you delete a row, all the rows after it get shifted up
one, so the index ints become stale and incorrect, causing Dee to
raise an unhandled exception about the row index being incorrect next
time you try to access a stale int. So I've thrown in a small check to
bump up the cached index ints each time a row is deleted.