Rip out ConnectionMonitor and Signaler, since they no longer make sense.
ConnectionMonitor existed mostly only to wrap the NetworkManager API,
to isolate us against incompatibilities between NM 0.8 and NM 0.9...
however, it turns out that Ubuntu 10.04 was the last Ubuntu release to
ship with 0.8. Additionally, since all it does is listen for NM events
over DBus, it doesn't actually do much of anything while
friends-service is not running; so this means that Ken's new vala
daemon is going to have to take over that role. It'll have to be smart
enough not to invoke our dbus methods during times that there is no
internet connection present.

As for Signaler, well, that was a very elegant design by Barry Warsaw,
but we were only using it to signal between ConnectionMonitor and the
Dispatcher (eg, to avoid unnecessary dbus round-trips), so there is no
longer any point in keeping that around.

This commit also adds a new dbus method onto our Dispatcher object,
for the purposes of downloading new contacts from all connected
protocols. This will be necessary in order for our new vala overlords
to invoke it over dbus periodically.