Avoid a race condition during shutdown.
Now, I don't have any evidence for this, but I had a gut feeling that
this shutdown logic might result in a race condition: the check for
whether this is the last thread to exit happens a little bit before
the thread actually exits, so it might theoretically have been
possible for two threads to exit near-simultaneously, each one seeing
that the other is still running, and then neither opting to initiate
the shutdown procedures.

This way, each thread termination is atomic: each thread is guaranteed
to check whether it is the last one, then perform it's shutdown logic,
without any other threads getting in between the check and the
termination.