Fix a bug with empty lists at Identica.
I started getting a traceback from Identica that we weren't getting
valid JSON from the service. Turns out I had no private messages, so
the json for that was '[]', but identica also wasn't supplying the
charset, so we needed at least 4 bytes in order to be able to
determine the charset.

In this case I decided to simply assume that the charset would be
utf8, since that's the most common, safest bet. Now it doesn't explode
when the server returns a JSON response of '[]'.