Move the Avatar expiry logic into service.vala rather than doing it in Python.
This was necessary because Avatar expiry is expensive (scanning a
large number of files and then delete them if the mtime is too old),
and this was happening synchronously at startup, delaying the
dispatcher from bringing up the dbus interface, delaying every single
dbus call (it's a holdover from when python was a long running
process).

This commit introduces a new dbus method called ExpireAvatars that
allows service.vala to trigger the avatar expiry logic at regular
intervals, rather than every single time every dbus method gets
called. By default it happens 5m after the initial startup, and then
once per week thereafter. These times can be easily finessed later on
if they are found to be deficient.

This commit also stops calling self.mainloop.quit() at the beginning
of each synchronous method call, which was found to be resulting in a
racy condition where a method would try to be called, and then it
would shut off the dbus interface before performing the work; this
meant that there were long-ish lengths of time where the dispatcher
couldn't be started, but also wouldn't respond to dbus method calls.
Instead, we use GLib.idle_add to call mainloop.quit, resulting in a
longer period of time in which the dispatcher's dbus methods can be
invoked before shutting down completely.