Give up on the idea of deduplication across protocols.
We used to have a *ton* of corner case logic that attempted to merge
messages that had the same content, but were from distinct social
networks. That made some sense back in the old gwibber days but it's
becoming an increasingly irrelevant corner case that is needlessly
complicating a lot of our application logic.

This commit removes that logic, and also greatly simplifies the model
schema accordingly, allowing for much easier filtering and sorting of
the model by it's consumers (such as qml-friends).

There's a big impact on the test suites, as this changes *huge*
portions of our fundamental architectural assumptions. In the process
of updating it, I also discovered that Base._unpublish_all was
poorly/indirectly tested, so I wrote some better, explicit tests for it.