Dump phony testsuite data into TestModel in --test mode.
This uses a mixture of our existing flickr, facebook, and twitter
sample data, using account_id's 1, 2, and 3 (making it sortable and
filterable for testing purposes).

This also moves a lot of the --test logic into friends.tests.mocks,
because it was growing a little bit out of hand in main.py and also
this made it easier to implement some of the mocks.