Fix the publishing of Facebook Stories (LP: #1155785)
The Facebook API docs were not explicit on the matter of whether a
Post object could ever contain both a "message" and a "story" at the
same time. However, it does say[0] that a story is "not intentionally
generated by users", which implies to me that a "story" could never
have a "message", and vice versa. If the user explicitly types a
message, that shows up under "message", but if the user just interacts
with some app or other (or accepts a friend request or that kind of
non-message activity), then that information will show up under
"story". So the code that I've written assumes that these two things
are mutually exclusive inside any given post. With excellent test
coverage, of course.

[0]: https://developers.facebook.com/docs/reference/api/post/