Remove unused methods from AccountManager.
It turns out that, like the recent purge-accounts logic that had to be
moved into service.vala, the AccountManager is also incapable of
responding to the enabled-event signal, because the python dispatcher
does not stick around in memory to receive it. Fortunately, this did
not result in a bug, because the account simply shows up as normal at
the next refresh and gets refreshed. So I've removed that unused code
and streamlined AccountManager.__init__ somewhat.

I believe that the AccountManager class is actually entirely
unnecessary at this point, and the remaining functionality it provides
can either be dropped, or merged into the Dispatcher. I'm still in the
process of determining where to draw that line, however.