Unify the Dispatcher shutdown logic.
Previously it was split between the Dispatcher and _OperationThread,
depending on whether the method was synchronous or not.

Now, I've created a decorator contextmanager that ensures the
Dispatcher stays alive for 30s beyond the return of the final method
call. Checks against quitting in the middle of threads, too. With test coverage.