Backport Raymond Hettinger's contextlib.ignored from Python 3.4.
This is compatible with Python 3.2 and 3.3, so is compatible with
quantal and raring, it just won't become part of the Python stdlib
until 3.4, which probably won't land in ubuntu until 14.04.

I've long considered "try: foo(); except Exception: pass" to be an ugly wart of
python, and this just slices it all away. This means it now only takes
one line instead of three when all you want to do is ignore an
exception.