Implement Base._set_cell, _inc_cell, and _dec_cell.
These provide a well-tested, centralized place for protocols to
manipulate the DeeModel without needing a reference to the model
itself.

This commit modifies Facebook.like and Facebook.unlike in order to
start using these, so that the model is correctly updated when posts
are liked or unliked. Also, many new tests are introduced to ensure
these features are working.