Rip out avatar expiry logic, and put real URLs into the model schema.
It turns out QML has it's own URL image cacher that caches images in
memory, which is way faster than the slow disk cache we had
implemented. So by disabling our own avatar cache, we're effecting a
large performance improvement in the QML frontend.

The only trick is, the Notify library doesn't have the same caching
abilities, so we do need to cache these images to disk just long
enough to show the notification bubble. But once that is displayed, we
no longer need to know or care what happens to the cache, so moving it
into /tmp makes it easier for the system to clean it up for us.

This commit also updates all tests to the new behavior as well.