Make contacts directly.
It turns out that EBookContacts.Contact is a subclass of
EBookContacts.VCard. This was a surprise to me, because the
documentation for it is nowhere to be found. I discovered that our
contact code creates a VCard, adds some attributes to it, and then
serializes it as a string, and then parses the string into a Contact
object. This is however, entirely unnecessary because we can just
create the Contact object directly, add vcard attributes to it, and be
done with it, thanks to inheritance.