Fix & re-enable Twitter test.
Let me be frank: This commit terrifies me. Implementing Twitter's
OAuth was the single most difficult and error-prone experience in all
of Friends. This test was specifically written with the intention that
it is the first line of defence against commits that send us back into
OAuth hell. I nearly cried when this test first started failing.

However, manual testing indicates that our oauth implementation is
working fine, just this test is failing for no discernable reason. So,
with that in mind, I'm changing the expected oauth signature to the
one that we are getting, effectively admitting that this was a false
positive all along.